backend/graph/requirements.txt
httpx==0.25.0
beautifulsoup4==4.12.2
neo4j==5.14.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2